Improved login

+++ moneymoney/views.py
@@ -76,21 +76,26 @@ class MyDjangoJSONEncoder(DjangoJSONEncoder):
 
 @api_view(['POST'])
 def login(request):
-    try:
-        user=User.objects.get(username=request.POST.get("username"))
-    except User.DoesNotExist:
-        return Response("Invalid user")
-        
-    password=request.POST.get("password")
-    pwd_valid=check_password(password, user.password)
-    if not pwd_valid:
-        return Response("Wrong password")
+    username=RequestString(request, "username")
+    password=RequestString(request, "password")
+    
+    if all_args_are_not_none(username, password):
+        try:
+            user=User.objects.get(username=username)
+        except User.DoesNotExist:
+            return Response("Invalid user")
+            
+        pwd_valid=check_password(password, user.password)
+        if not pwd_valid:
+            return Response("Wrong password")
 
-    if Token.objects.filter(user=user).exists():#Lo borra
-        token=Token.objects.get(user=user)
-        token.delete()
-    token=Token.objects.create(user=user)
-    return Response(token.key)
+        if Token.objects.filter(user=user).exists():#Lo borra
+            token=Token.objects.get(user=user)
+            token.delete()
+        token=Token.objects.create(user=user)
+        return Response(token.key)
+    else:
+        return Response(_("Bad credentials"))
     
 @api_view(['POST'])
 def logout(request):
